Standalone mock API Auth
