+++ packages/admin-ui/src/pages/system/components/Profile/ChangePassword.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import Button from "components/Button";
 import ErrorView from "components/ErrorView";
-import { fetchChangePass, fetchLogoutAndReload } from "api/auth";
+import { authApi } from "api/auth";
 import { ReqStatus } from "types";
 import Ok from "components/Ok";
 import {
@@ -29,14 +29,14 @@ export function ChangePassword() {
     if (isValid)
       try {
         setReqStatus({ loading: true });
-        await fetchChangePass({
+        await authApi.changePass({
           password: oldPassword,
           newPassword
         });
         setReqStatus({ result: true });
 
         // Logout so user re-enters the password
-        await fetchLogoutAndReload();
+        await authApi.logoutAndReload();
       } catch (e) {
         setReqStatus({ error: e });
       }
